requirements.txt
Django==4.0.6
django_countries==7.3.2
django_rest_knox==4.2.0
django_rest_passwordreset==1.2.1
djangorestframework==3.13.1
knox==0.1.14
